Create Login and Switch provider components
These components are styling independant so apps can use them if the default button components don't cover the needed use cases.
