+++ addon/components/acmidm-login.js
@@ -1,31 +1,4 @@
-import { warn } from '@ember/debug';
-import Component from '@glimmer/component';
-import { inject as service } from '@ember/service';
-import { tracked } from '@glimmer/tracking';
-import { action } from '@ember/object';
+import templateOnly from '@ember/component/template-only';
 
-export default class AcmIdmLoginComponent extends Component {
-  @tracked errorMessage = '';
-  @tracked isAuthenticating = false;
-  @service session;
-
-  @action
-  login() {
-    this.errorMessage = '';
-    this.isAuthenticating = true;
-    this.session
-      .authenticate('authenticator:torii', 'acmidm-oauth2')
-      .catch((reason) => {
-        warn(reason.error || reason, { id: 'authentication.failure' });
-
-        if (reason.status == 403)
-          this.errorMessage = 'U heeft geen toegang tot deze applicatie.';
-        else
-          this.errorMessage =
-            'Fout bij het aanmelden. Gelieve opnieuw te proberen.';
-      })
-      .finally(() => {
-        this.isAuthenticating = false;
-      });
-  }
-}
+const AcmidmLoginComponent = templateOnly();
+export default AcmidmLoginComponent;
