+++ addon/components/acmidm-switch.hbs
@@ -1,7 +1,9 @@
-<button ...attributes disabled={{this.disabled}} type="button" {{on "click" this.switch}}>
-  {{#if (has-block)}}
-  {{yield}}
-{{else}}
-  Wissel van bestuurseenheid
-{{/if}}
-</button>
+<Acmidm::Switch as |acmidm|>
+  <button ...attributes disabled={{acmidm.isSwitching}} type="button" {{on "click" acmidm.switch}}>
+    {{#if (has-block)}}
+      {{yield}}
+    {{else}}
+      Wissel van bestuurseenheid
+    {{/if}}
+  </button>
+</Acmidm::Switch>
