+++ addon/components/acmidm/login.hbs
@@ -1,0 +1,5 @@
+{{yield (hash
+  login=this.login
+  isAuthenticating=this.isAuthenticating
+  errorMessage=this.errorMessage
+)}}
