+++ addon/components/acmidm/switch.hbs
@@ -1,0 +1,4 @@
+{{yield (hash
+  switch=this.switch
+  isSwitching=this.isSwitching
+)}}
