+++ app/components/acmidm/login.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-acmidm-login/components/acmidm/login';
