+++ app/components/acmidm/switch.js
@@ -1,0 +1,1 @@
+export { default } from '@lblod/ember-acmidm-login/components/acmidm/switch';
